requirements/base.txt
# CORE
Django==4.1.2

# DRF
drf-yasg==1.21.4
django-filter==22.1
drf-writable-nested==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

# Configuration
django-split-settings==1.2.0

# DB
django-redis==5.2.0

# Utilities
persian-tools==0.0.10

# Celery
celery==5.2.7
django-celery-beat==2.4.0

# Notifications
kavenegar==1.1.2

# Test
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
